Abstract carousel for destinations
Create a shared view listener for the destination carousels on the home view and the destination details view.
